requirements.txt
Flask==2.3.3
requests==2.31.0
python-dotenv==1.1.1
APScheduler==3.11.0
Flask-APScheduler==1.13.1
